+++ json.go
@@ -1,0 +1,34 @@
+package main
+
+import (
+	"log"
+	"encoding/json"
+	"io/ioutil"
+)
+
+var articles Article
+
+func getJSON()([]byte, error) {
+  resp, err := getResponse()
+
+  byteArray, err := ioutil.ReadAll(resp.Body)
+  if err != nil {
+    return nil, err
+  }
+  defer resp.Body.Close()
+
+  return byteArray, err
+}
+
+func encodeJSONToGo()(Article, error) {
+
+  byteArray, err := getJSON()
+  if err != nil {
+    log.Fatalf("Error!: %v", err)
+  }
+
+  if err := json.Unmarshal(byteArray, &articles); err != nil {
+    log.Fatalf("Error!: %v", err)
+  }
+  return articles, err
+}
